Add departures to routes API endpoint
Departures are nested inside stops in the routes endpoint. To get them
returned, one must specify the date using query param
"date=yyyy-mm-dd". One can also filter the departures by direction
using query param "direction_id=<0 or 1>".
